Added new Features

+++ customer/src/main/java/com/quantumai/customer/controller/UsersAPI.java
@@ -1,10 +1,14 @@
 package com.quantumai.customer.controller;
+
 import com.quantumai.customer.dto.*;
 import com.quantumai.customer.entity.*;
+import com.quantumai.customer.exception.NoSubscriptionError;
 import com.quantumai.customer.exception.UserException;
 import com.quantumai.customer.repository.CustomRoleRepository;
 import com.quantumai.customer.repository.CustomerRepository;
+import com.quantumai.customer.repository.SubscriptionRepository;
 import com.quantumai.customer.repository.UsersRepository;
+import com.quantumai.customer.security.JwtService;
 import com.quantumai.customer.service.UserService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -42,20 +46,44 @@ public class UsersAPI {
 
   @Autowired private CustomRoleRepository customRoleRepository;
 
-  @PostMapping("/registerUser")
-  public void register(@RequestBody Users users) throws UserException {
+  @Autowired private JwtService jwtService;
 
+  @Autowired SubscriptionRepository subscriptionRepository;
+
+  @PostMapping("/registerUser")
+  public void register(@RequestBody Users users, @RequestHeader String Authorization)
+      throws UserException, NoSubscriptionError {
+    //      users.setStatus(StatusEnum.active);
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.registerUser(users);
   }
 
   @PostMapping("/send/{companyId}")
-  public void sendEmail(@PathVariable String companyId, @RequestBody Mail mail) {
+  public void sendEmail(@PathVariable Long companyId, @RequestBody Mail mail)
+      throws NoSubscriptionError {
+
+    Optional<Subscription> subscriptionOptional =
+        subscriptionRepository.findByCompanyIdAndStatus(companyId, "ACTIVE");
+    if (subscriptionOptional.isEmpty()) {
+      throw new NoSubscriptionError("No Active Subscription");
+    }
+
     AuthenticationResponseDTO response = userService.generateToken(mail);
     System.out.print(
         "---------------------///////--------------------////-----------------"
             + response.getToken());
-//    String emailId = "http://localhost:4200/invitation/";
-    		String emailId="http://assetyugg.com.s3-website-us-east-1.amazonaws.com/invitation/";
+    //    String emailId = "http://localhost:4200/invitation/";
+    String emailId = "http://assetyugg.com.s3-website-us-east-1.amazonaws.com/invitation/";
     userService.sendSimpleMessage(
         mail.getEmail(),
         "Invitation mail: ",
@@ -63,7 +91,7 @@ public class UsersAPI {
   }
 
   //	@GetMapping(value="/getCustomer/{companyId}")
-  //	public ResponseEntity<Customer> getCustomerDetails(@PathVariable String companyId,
+  //	public ResponseEntity<Customer> getCustomerDetails(@PathVariable Long companyId,
   // @RequestHeader("Authorization") String token){
   //
   //		HttpHeaders headers = new HttpHeaders();
@@ -81,8 +109,19 @@ public class UsersAPI {
   //
   //	}
   @PostMapping(value = "/invite")
-  public void getInviteDetails(@RequestBody UsersDTO userDTO) throws UserException {
-    //		Claims myDetails=userService.decodeDetails(details);
+  public void getInviteDetails(@RequestBody UsersDTO userDTO, @RequestHeader String Authorization)
+      throws UserException, NoSubscriptionError {
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     System.out.println(userDTO);
     Users user = modelMapper.map(userDTO, Users.class);
     //		user.setPassword(passwordEncoder.encode(user.getPassword()));
@@ -93,25 +132,26 @@ public class UsersAPI {
     //		else if(myDetails.get("role").equals("TECHNICAL")) {
     //			user.setRole(Role.TECHNICAL);
     //		}
+    user.setStatus(StatusEnum.inActive);
     userService.registerUser(user);
   }
 
   @GetMapping(value = "/getUsers/{companyId}")
-  public ResponseEntity<List<UsersDTO>> getAllUsers(@PathVariable String companyId) {
+  public ResponseEntity<List<UsersDTO>> getAllUsers(@PathVariable Long companyId) {
     List<UsersDTO> userList = userService.getAllUsers(companyId);
     return ResponseEntity.ok(userList);
   }
 
   @GetMapping(value = "/invite/getUser/{companyId}/{details}")
   public ResponseEntity<UsersDTO> getUsers(
-      @PathVariable String companyId, @PathVariable String details) {
+      @PathVariable Long companyId, @PathVariable String details) {
     Claims myDetails = userService.decodeDetails(details);
     UsersDTO user = userService.getUsers(companyId, myDetails.get("email").toString());
     return ResponseEntity.ok(user);
   }
 
   @GetMapping(value = "/getTechnicalUser/{companyId}")
-  public ResponseEntity<List<UsersDTO>> getTechnicalUser(@PathVariable String companyId) {
+  public ResponseEntity<List<UsersDTO>> getTechnicalUser(@PathVariable Long companyId) {
 
     List<UsersDTO> userDTOList = userService.getAllUsersByRole("TECHNICAL", companyId);
     return ResponseEntity.ok(userDTOList);
@@ -119,16 +159,48 @@ public class UsersAPI {
 
   @GetMapping(value = "/getUserDetails/{companyId}/{email}")
   public ResponseEntity<UsersDTO> getUserDetails(
-      @PathVariable String companyId, @PathVariable String email) {
+      @PathVariable Long companyId, @PathVariable String email) {
 
     UsersDTO UserDTO = userService.getUsers(companyId, email);
     return ResponseEntity.ok(UserDTO);
   }
 
+  @PutMapping(value = "/userDetails")
+  public ResponseEntity<String> updateUserDetails(
+      @RequestBody UsersDTO usersDTO, @RequestHeader String Authorization)
+      throws NoSubscriptionError {
+
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
+    userService.updateUser(usersDTO);
+    return ResponseEntity.ok("Status Updated Successfully");
+  }
+
+  //    @Transactional
   @PostMapping(value = "/updateUserDetails")
   public void updateUserDetails(@RequestBody UsersDTO usersDTO, HttpServletRequest request)
       throws Exception {
     String authorizationHeader = request.getHeader("Authorization");
+    String jwt = authorizationHeader.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     if (authorizationHeader == null || !authorizationHeader.startsWith("Bearer ")) {
       throw new RuntimeException("Invalid or missing Authorization header");
     }
@@ -171,6 +243,16 @@ public class UsersAPI {
           throw new Exception("Unkown Role");
         }
       }
+      Optional<Customer> customerOptional = customerRepository.findByEmail(usersDTO.getEmail());
+      if (customerOptional.isPresent()) {
+        if (usersDTO.getRole() != null) {
+          customerOptional.get().setRole(usersDTO.getRole().getName());
+        }
+        customerOptional.get().setFirstName(usersDTO.getFirstName());
+        customerOptional.get().setLastName(usersDTO.getLastName());
+        customerOptional.get().setMobileNumber(usersDTO.getMobileNumber());
+        customerRepository.save(customerOptional.get());
+      }
 
     } else {
       throw new Exception("Unkown Email");
@@ -179,16 +261,40 @@ public class UsersAPI {
 
   @DeleteMapping(value = "/deleteUserDetails/{companyId}/{email}")
   public void deleteUserDetails(
-      @PathVariable String companyId,
+      @PathVariable Long companyId,
       @PathVariable String email,
       @RequestHeader("Authorization") String authHeader)
       throws Exception {
     //		System.out.print("-,,,,---,,,,---"+authHeader);
+    String jwt = authHeader.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.deleteUser(companyId, email, authHeader);
   }
 
   @PostMapping("/addfields")
-  public void addNewFields(@RequestBody UserExtraFieldsDTO extraFieldsDTO) throws Exception {
+  public void addNewFields(
+      @RequestBody UserExtraFieldsDTO extraFieldsDTO, @RequestHeader String Authorization)
+      throws Exception {
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.addExtraFields(extraFieldsDTO);
   }
 
@@ -198,45 +304,108 @@ public class UsersAPI {
   }
 
   @DeleteMapping("/deleteExtraFields/{id}")
-  public void deleteExtraField(@PathVariable String id) throws Exception {
+  public void deleteExtraField(@PathVariable String id, @RequestHeader String Authorization)
+      throws Exception {
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.deleteExtraFields(id);
   }
 
   @GetMapping("/getExtraFieldName/{companyId}")
-  public List<UserExtraFieldNameDTO> getExtraFieldName(@PathVariable String companyId) {
+  public List<UserExtraFieldNameDTO> getExtraFieldName(@PathVariable Long companyId) {
     // System.out.println("----------my company------------->"+companyId);
     return userService.getUserExtraField(companyId);
   }
 
   @PostMapping("/addExtraFieldName")
-  public void addExtraFieldName(@RequestBody UserExtraFieldNameDTO extraFieldNameDTO)
-          throws Exception {
+  public void addExtraFieldName(
+      @RequestBody UserExtraFieldNameDTO extraFieldNameDTO, @RequestHeader String Authorization)
+      throws Exception {
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.addUserExtraField(extraFieldNameDTO);
   }
 
   @DeleteMapping("/deleteExtraFieldName/{id}")
-  public void deleteExtraFieldName(@PathVariable String id) {
+  public void deleteExtraFieldName(@PathVariable String id, @RequestHeader String Authorization)
+      throws NoSubscriptionError {
     // System.out.println("-----------------------api------------------------>"+id);
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.deleteUserExtraField(id);
   }
 
   @GetMapping("/getExtraFieldNameValue/{companyId}")
-  public Map<String, Map<String, String>> getExtraFieldNameValue(@PathVariable String companyId) {
+  public Map<String, Map<String, String>> getExtraFieldNameValue(@PathVariable Long companyId) {
     return userService.getextraFieldList(companyId);
   }
+
   @PostMapping("/mandatoryFields")
-  public void mandatoryFields(@RequestBody UserMandatoryFields mandatoryFields) {
+  public void mandatoryFields(
+      @RequestBody UserMandatoryFields mandatoryFields, @RequestHeader String Authorization)
+      throws NoSubscriptionError {
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.updateMandatoryFields(mandatoryFields);
   }
 
   @PostMapping("/showFields")
-  public void showFields(@RequestBody UserShowFields showFields) {
+  public void showFields(
+      @RequestBody UserShowFields showFields, @RequestHeader String Authorization)
+      throws NoSubscriptionError {
+    String jwt = Authorization.substring(7);
+    String userEmail = jwtService.extractUserEmail(jwt);
+    Optional<Customer> customer = customerRepository.findByEmail(userEmail);
+    if (customer.isPresent()) {
+
+      Optional<Subscription> subscriptionOptional =
+          subscriptionRepository.findByCompanyIdAndStatus(customer.get().getCompanyId(), "ACTIVE");
+      if (subscriptionOptional.isEmpty()) {
+        throw new NoSubscriptionError("No Active Subscription");
+      }
+    }
     userService.updateShowFields(showFields);
   }
 
   @GetMapping("/getMandatoryFields/{name}/{companyId}")
   public ResponseEntity<UserMandatoryFields> getMandatoryFields(
-          @PathVariable String name, @PathVariable String companyId) {
+      @PathVariable String name, @PathVariable Long companyId) {
     // System.out.println("============================>"+name+companyId);
     UserMandatoryFields mandatoryFields = userService.getMandatoryFields(name, companyId);
     return ResponseEntity.ok(mandatoryFields);
@@ -244,26 +413,34 @@ public class UsersAPI {
 
   @GetMapping("/getShowFields/{name}/{companyId}")
   public ResponseEntity<UserShowFields> getShowFields(
-          @PathVariable String name, @PathVariable String companyId) {
+      @PathVariable String name, @PathVariable Long companyId) {
     UserShowFields showFields = userService.getShowFields(name, companyId);
     return ResponseEntity.ok(showFields);
   }
 
   @GetMapping("/getAllMandatoryFields/{companyId}")
   public ResponseEntity<List<UserMandatoryFields>> getAllMandatoryFields(
-          @PathVariable String companyId) {
+      @PathVariable Long companyId) {
     List<UserMandatoryFields> mandatoryFieldsList = userService.getAllMandatoryFields(companyId);
     return ResponseEntity.ok(mandatoryFieldsList);
   }
 
   @GetMapping("/getAllShowFields/{companyId}")
-  public ResponseEntity<List<UserShowFields>> getAllShowFields(@PathVariable String companyId) {
+  public ResponseEntity<List<UserShowFields>> getAllShowFields(@PathVariable Long companyId) {
     List<UserShowFields> showFieldsList = userService.getAllShowFields(companyId);
     return ResponseEntity.ok(showFieldsList);
   }
 
   @DeleteMapping("/deleteShowAndMandatoryField/{name}/{companyId}")
-  public void showFields(@PathVariable String name, @PathVariable String companyId) {
+  public void showFields(@PathVariable String name, @PathVariable Long companyId)
+      throws NoSubscriptionError {
+
+    Optional<Subscription> subscriptionOptional =
+        subscriptionRepository.findByCompanyIdAndStatus(companyId, "ACTIVE");
+    if (subscriptionOptional.isEmpty()) {
+      throw new NoSubscriptionError("No Active Subscription");
+    }
+
     userService.deleteShowAndMandatoryFields(companyId, name);
   }
 }
